Refactor lex.go
Split lex.go into several unit files, in preparation for testing
session.

Improve the Lexer type to more closely follow the immutability
principles.

Improve the Position type to more closely follow the immutability
principle.

Unexport types, variables and methods that are not meant to be used
outside of the package.