Refactoring
In composite_lexer, ensure that the literal is the whole operator.

Split the dispatcher_lexer to follow the Single-Responsibility
Principle, into eof_lexer, eol_lexer and space_lexer.

Create Type(), Position() and Literal() getters for the Token type. Use
them when needed in the implementation.

Set nextLexerFunc to eofLexer in New()